requirements.txt
fastapi==0.123.9
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
anthropic==0.40.0
sentence-transformers==2.7.0
numpy==2.1.3
pydantic==2.9.2
sqlalchemy==2.0.44
python-multipart==0.0.12

